Making the appendix only 30 pages & removing some data
This commit forks the branch of the replication package.  I have gone through and made formatting changes and very small changes to the appendix contents (namely removing a section of analysis from the attitudes appendix that we never reference elsewhere).  In addition, I have removed from the source data and intermediate data files a few files that we do not use and we do not (I believe) have permission to share.